+++ code/ext/ext.lib.automerge.webrtc/src/ui/ui.Sample.02/ui.Loader.tsx
@@ -1,4 +1,4 @@
-import { useLoader } from './Load.useLoader';
+import { useLoader } from './ui.useLoader';
 import { COLORS, Color, Spinner, css, type t } from './common';
 
 export type LoaderProps = {
